# statshouse-ui/public/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StatsHouse - OpenAPI 3.0",
    "description": "",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/metric": {
      "get": {
        "description": "Get metric meta info",
        "parameters": [
          {
            "$ref": "#/components/parameters/paramMetric"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MetricInfo"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Edit metric meta info",
        "parameters": [
          {
            "$ref": "#/components/parameters/paramMetric"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetricInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/MetricInfo"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/metric-tag-values": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/paramMetric"
          },
          {
            "$ref": "#/components/parameters/paramVersion"
          },
          {
            "$ref": "#/components/parameters/paramNumResult"
          },
          {
            "$ref": "#/components/parameters/paramTagID"
          },
          {
            "$ref": "#/components/parameters/paramFrom"
          },
          {
            "$ref": "#/components/parameters/paramTo"
          },
          {
            "$ref": "#/components/parameters/paramWhat"
          },
          {
            "$ref": "#/components/parameters/paramFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GetMetricTagValuesResp"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "paramMetric": {
        "name": "s",
        "in": "query",
        "description": "Name metric",
        "required": true,
        "allowEmptyValue": false,
        "schema": {
          "type": "string"
        },
        "example": ""
      },
...
      "paramFrom": {
        "name": "f",
        "in": "query",
        "description": "from time",
        "required": true,
        "allowEmptyValue": false,
        "schema": {
          "type": "number"
        },
        "example": 1665585339
      },
...
      "paramTo": {
        "name": "t",
        "in": "query",
        "description": "to time",
        "required": true,
        "allowEmptyValue": false,
        "schema": {
          "type": "number"
        },
        "example": 1665758140
      },
...
      "paramWhat": {
        "name": "qw",
        "in": "query",
        "description": "what",
        "required": true,
        "allowEmptyValue": false,
        "schema": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/queryFn"
          }
        },
        "example": [
          "count_norm"
        ]
      },
...
      "paramVersion": {
        "name": "v",
        "in": "query",
        "description": "backend version",
        "required": true,
        "allowEmptyValue": false,
        "schema": {
          "type": "string",
          "enum": [
            "1",
            "2"
          ]
        },
        "example": 2
      },
      "paramNumResult": {
        "name": "n",
        "in": "query",
        "description": "num results",
        "required": true,
        "allowEmptyValue": false,
        "schema": {
          "type": "number"
        },
        "example": 5
      },
      "paramTagID": {
        "name": "k",
        "in": "query",
        "description": "Tag ID",
        "required": true,
        "allowEmptyValue": false,
        "schema": {
          "type": "number"
        },
        "example": 0
      },
      "paramFilter": {
        "name": "qf",
        "in": "query",
        "description": "Filter",
        "required": false,
        "allowEmptyValue": false,
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
...
    },
    "schemas": {
      "MetricKind": {
        "type": "string",
        "enum": [
          "counter",
          "value",
          "value_p",
          "unique",
          "mixed",
          "mixed_p"
        ]
      },
      "MetricMetaTagRawKind": {
        "type": "string",
        "enum": [
          "uint",
          "hex",
          "hex_bswap",
          "timestamp",
          "timestamp_local",
          "ip",
          "ip_bswap"
        ]
      },
      "MetricMetaTag": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "raw": {
            "type": "boolean"
          },
          "raw_kind": {
            "$ref": "#/components/schemas/MetricMetaTagRawKind"
          },
          "id2value": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "value_comments": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "MetricMetaValue": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "metric_id": {
            "type": "number"
          },
          "namespace_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetricMetaTag"
            }
          },
          "visible": {
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/MetricKind"
          },
          "weight": {
            "type": "number"
          },
          "resolution": {
            "type": "number"
          },
          "string_top_name": {
            "type": "string"
          },
          "string_top_description": {
            "type": "string"
          },
          "pre_key_tag_id": {
            "type": "string"
          },
          "pre_key_from": {
            "type": "number"
          },
          "update_time": {
            "type": "number"
          },
          "version": {
            "type": "number"
          },
          "group_id": {
            "type": "number"
          },
          "pre_key_only": {
            "type": "boolean"
          },
          "metric_type": {
            "type": "string"
          },
          "tags_draft": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/MetricMetaTag"
            }
          }
        },
        "required": [
          "name",
          "kind",
          "metric_type",
          "namespace_id"
        ]
      },
      "MetricInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "metric": {
            "$ref": "#/components/schemas/MetricMetaValue"
          }
        },
        "required": [
          "metric"
        ]
      },
...
      "queryFn": {
        "type": "string",
        "enum": [
          "count",
          "count_norm",
          "cu_count",
          "min",
          "max",
          "avg",
          "cu_avg",
          "sum",
          "sum_norm",
          "cu_sum",
          "stddev",
          "p25",
          "p50",
          "p75",
          "p90",
          "p95",
          "p99",
          "p999",
          "unique",
          "unique_norm",
          "max_host",
          "max_count_host",
          "dv_count",
          "dv_sum",
          "dv_avg",
          "dv_min",
          "dv_max",
          "dv_unique",
          "dv_unique_norm",
          "dv_count_norm",
          "dv_sum_norm"
        ]
      },
...
      "MetricTagValueInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "value": {
            "type": "string"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "value",
          "count"
        ]
      },
      "GetMetricTagValuesResp": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "tag_values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetricTagValueInfo"
            }
          },
          "tag_values_more": {
            "type": "boolean"
          }
        },
        "required": [
          "tag_values",
          "tag_values_more"
        ]
      },
...
    },
...
  }
}